backend/requirements.txt
# MAIDAR Backend Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.36  # Updated for Python 3.13 support
alembic==1.14.0  # Updated for compatibility
psycopg[binary]==3.2.3  # PostgreSQL adapter (psycopg3 with Python 3.13 support)
asyncpg==0.30.0  # Async PostgreSQL - Updated for Python 3.13

# Data Validation & Serialization
pydantic==2.10.5  # Updated for Python 3.13 support
pydantic-settings==2.7.1  # Updated for Python 3.13 support

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7

# Redis (caching)
redis==5.2.1  # Updated for Python 3.13
# hiredis==3.0.0  # Commented out - requires C++ build tools, optional performance enhancement

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing FastAPI

# Development Tools
black==23.12.0  # Code formatting
flake8==6.1.0  # Linting
mypy==1.7.1  # Type checking
isort==5.13.2  # Import sorting

# API Documentation
fastapi[all]==0.104.1

# Date/Time handling
python-dateutil==2.8.2

# Environment & Configuration
pyyaml==6.0.2  # Updated for Python 3.13

# Monitoring & Logging (optional for production)
# sentry-sdk==1.39.1

# AWS SDK (if deploying to AWS UAE region)
# boto3==1.34.0
